+++ src/test/java/domain/WinningLottoTest.java
@@ -1,14 +1,13 @@
 package domain;
 
+import static exception.ExceptionMessage.*;
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 import constant.WinningCount;
-import exception.ErrorMessage;
+import exception.ExceptionMessage;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -27,7 +26,7 @@ class WinningLottoTest {
         // when & then
         assertThatThrownBy(()-> new WinningLotto(numbers,bonusNumber))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(ErrorMessage.LOTTO_RANGE_ERROR.getMessage());
+                .hasMessage(LOTTO_RANGE_ERROR.getMessage());
     }
 
     @DisplayName("보너스 번호 중복 검증")
@@ -39,7 +38,7 @@ class WinningLottoTest {
         // when & then
         assertThatThrownBy(()-> new WinningLotto(numbers,bonusNumber))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(ErrorMessage.LOTTO_NUMBER_DUPLICATED_ERROR.getMessage());
+                .hasMessage(LOTTO_NUMBER_DUPLICATED_ERROR.getMessage());
     }
 
     @DisplayName("로또 결과 반환 검증")
